Align auth layouts and finalize SQLite-backed reservation/order APIs.
Center login/register/account logged-out UI, add DatabaseHelper int-based overloads plus transactional insertOrder, and sync menu/requests/order fragments to reload real per-user data from session-aware DB flows.

+++ app/src/main/java/com/example/quanlynhahang/OrderFragment.java
@@ -86,6 +86,6 @@ public class OrderFragment extends Fragment {
             return;
         }
 
-        orders.addAll(databaseHelper.getOrdersByUserId(userId));
+        orders.addAll(databaseHelper.getOrdersByUserId((int) userId));
     }
 }
